runtime/requirements.txt
aiofiles==0.5.0
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.20.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
databases==0.3.2
fastjsonschema==2.14.4
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
idna==2.9
pkg-resources==0.0.0
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.13.0
python-multipart==0.0.5
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.17
starlette==0.13.4
toml==0.10.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
